refactor(#1338): canonicalize in-package imports off shim paths
Replaces shim-path imports inside `src/icom_lan/<layer>/*.py` with
canonical layer paths so first-party code stops routing through the
sys.modules-alias re-export shims. Behavior is identical at runtime
(the shim and canonical names resolve to the same module objects),
but canonical paths match the stated layer-boundary contract in
CLAUDE.md and let mkdocstrings/griffe traverse references reliably.

Mechanical search-and-replace across 22 files (~61 lines changed):
no imports added or removed, no reordering, no behavior changes,
shim files at the package root untouched, public API surface
(`__init__.py` Tier 1/Tier 2 blocks) untouched, TYPE_CHECKING
blocks left as-is per the issue brief, tests/ untouched.

Verification:
- uv run lint-imports: 4 contracts kept, 0 broken
- uv run pytest tests/ ... --ignore=tests/integration: 5320 passed
- uv run ruff check src/ tests/: clean
- uv run mypy src/: clean

Closes follow-up commit-3 of Plan §5.3 Step 13.
