+++ src/icom_lan/commands/commander.py
@@ -6,8 +6,8 @@ from dataclasses import dataclass
 from enum import IntEnum
 from typing import TypeVar
 
-from icom_lan.exceptions import ConnectionError
-from icom_lan.types import CivFrame
+from icom_lan.core.exceptions import ConnectionError
+from icom_lan.core.types import CivFrame
 
 __all__ = ["IcomCommander", "Priority"]
