+++ src/icom_lan/profiles/rig_loader.py
@@ -9,11 +9,11 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Any
 
-from icom_lan.capabilities import KNOWN_CAPABILITIES
-from icom_lan.command_map import CommandMap
+from icom_lan.core.capabilities import KNOWN_CAPABILITIES
+from icom_lan.commands.command_map import CommandMap
 
 __all__ = ["RigConfig", "RigLoadError", "load_rig", "discover_rigs"]
-from icom_lan.command_spec import CatCommandSpec, CivCommandSpec, CommandSpec
+from icom_lan.commands.command_spec import CatCommandSpec, CivCommandSpec, CommandSpec
 from icom_lan.profiles import (
     BandInfo,
     ControlSpec,
