+++ src/icom_lan/rigctld/state_cache.py
@@ -1,4 +1,4 @@
 # rigctld/state_cache.py — backward-compat re-export
-from icom_lan._state_cache import CacheField, StateCache  # noqa: F401
+from icom_lan.core._state_cache import CacheField, StateCache  # noqa: F401
 
 __all__ = ["CacheField", "StateCache"]
