+++ src/icom_lan/runtime/_dual_rx_runtime.py
@@ -33,8 +33,8 @@ from icom_lan.commands import get_unselected_freq as _get_unselected_freq_cmd
 from icom_lan.commands import get_unselected_mode as _get_unselected_mode_cmd
 from icom_lan.commands import parse_selected_freq_response as _parse_selected_freq_response
 from icom_lan.commands import parse_selected_mode_response as _parse_selected_mode_response
-from icom_lan.exceptions import CommandError, TimeoutError
-from icom_lan.types import Mode
+from icom_lan.core.exceptions import CommandError, TimeoutError
+from icom_lan.core.types import Mode
 
 # CI-V command byte for VFO select / equal / swap (0x07).
 _CMD_VFO = 0x07
