+++ src/icom_lan/runtime/sync.py
@@ -21,8 +21,8 @@ from .ic705 import (
     restore_ic705_data_profile as _restore_ic705_data_profile,
 )
 from .radio import IcomRadio as _AsyncIcomRadio, _DEFAULT_AUDIO_CODEC  # noqa: TID251
-from icom_lan.capabilities import CAP_METERS, CAP_POWER_CONTROL
-from icom_lan.types import AudioCodec, Mode, ScopeCompletionPolicy
+from icom_lan.core.capabilities import CAP_METERS, CAP_POWER_CONTROL
+from icom_lan.core.types import AudioCodec, Mode, ScopeCompletionPolicy
 
 T = TypeVar("T")
